refactor: extract bending-tilt parameter helpers to bt_params.py
Extract low-risk configuration and parameter resolution logic from bending_tilt_leaflet.py to modules/energy/bt_params.py. Centralizes theory-mode and benchmark-scoped flag resolution while reducing core module size by ~200 lines. Verified with bending/tilt energy and regression tests.
